코프태그 작업

+++ src/main/java/com/itwill/ilhajob/common/dto/CorpTagDto.java
@@ -15,5 +15,5 @@ import lombok.ToString;
 public class CorpTagDto {
 	private long id;
 	private String corpId;
-	private int tagId;
+	private long tagId;
 }
